requirements.txt
python-dotenv==0.19.2
colorama==0.4.4